requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.1
discord==1.0.1
discord.py==1.2.5
dnspython==1.16.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.5.2
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==6.2.1
pylint==2.4.3
pymongo==3.9.0
requests==2.22.0
six==1.13.0
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
